Updated image retrieve function

+++ image_retrieve.py
@@ -13,29 +13,27 @@ def clean_folder(folder_path):
 
     for item in os.listdir(folder_path):
         item_path = os.path.join(folder_path, item)
-        try:
-            if os.path.isfile(item_path) or os.path.islink(item_path):
-                os.remove(item_path)
-                print(f"Deleted file: {item_path}")
-            elif os.path.isdir(item_path):
-                shutil.rmtree(item_path)
-                print(f"Deleted folder: {item_path}")
-        except Exception as e:
-            print(f"Error deleting {item_path}: {e}")
+        if item_path== os.path.join(folder_path, 'urls.txt'):
+            with open(item_path, 'w') as f:
+                f.write('')
+        else:
+            try:
+                if os.path.isfile(item_path) or os.path.islink(item_path):
+                    os.remove(item_path)
+                    print(f"Deleted file: {item_path}")
+                elif os.path.isdir(item_path):
+                    shutil.rmtree(item_path)
+                    print(f"Deleted folder: {item_path}")
+            except Exception as e:
+                print(f"Error deleting {item_path}: {e}")
 
 
-def image_search(query:str, max_results:int = 3):
-    results = DDGS().images(query=query, max_results=max_results)
-    return results
-
-
-def save_image(image_results, folder="downloads"):
-    clean_folder(folder)
-    os.makedirs(folder, exist_ok=True)
-
-    for i, result in enumerate(image_results):
-        url = result["image"]
-        try:
+def process_image(result, urls, saved_images, folder):
+    url = result["image"]
+    if url in urls:
+            return saved_images
+    
+    try:
             response = requests.get(url, timeout=10)
             response.raise_for_status()
 
@@ -44,17 +42,43 @@ def save_image(image_results, folder="downloads"):
             image_format = image.format.lower()  # jpg, png, webp, ecc.
 
             # Nome file con estensione corretta
-            file_name = f"image_{i+1}.{image_format}"
+            file_name = f"image_{saved_images+1}.{image_format}"
             path = os.path.join(folder, file_name)
 
             image.save(path)
             print(f"Saved: {path}")
 
-        except Exception as e:
+            with open(os.path.join(folder,'urls.txt'),'a') as f:
+                f.write(url+'\n')
+            urls.append(url)
+
+            saved_images+=1
+            return saved_images
+
+    except Exception as e:
             print(f"Error with {url}: {e}")
+            return saved_images
+    
 
+    
 def image_retrieve(query, max_results:int = 3, folder="downloads"):
-    save_image(image_search(query,max_results=max_results), folder=folder)
+    # save_image(image_search(query,max_results=max_results), folder=folder)
+    image_results = DDGS().images(query=query, max_results=max_results)
+    saved_images=0
+    
+    clean_folder(folder)
+    os.makedirs(folder, exist_ok=True)
+    with open(os.path.join(folder,'urls.txt'),'r') as f:
+        urls= f.readlines()
+    
+    for result in image_results:
+        saved_images= process_image(result, urls, saved_images, folder)
+    
+    while saved_images < max_results:
+        image_results = DDGS().images(query=query, max_results=max_results-saved_images+1)
+        for result in image_results:
+            saved_images= process_image(result, urls, saved_images, folder)
 
 
-image_retrieve('example')
+
+image_retrieve('example', max_results=20)
